Eliminate initial flash of size by looking for the window width immediately instead of in a callback
